feat(resources): generate v1.1 PDFs + manifest
- Wire up resources page to use manifest-only (no fallback)
- Generate HTML resources with full content + proper styling
- Create organized category structure in public/resources/
- Update package.json with gen:resources script
- Fix encoding issues with UTF-8 and proper font loading
- Include Düsseldorf footer and A4 layout in all resources
- Resources v1.1 content: parent-email-playbook, assessment-rubrics-pack,
  report-comment-bank, lesson-plan-template-primary
