+++ package.json
@@ -15,7 +15,7 @@
                     "postbuild":  "echo \u0027Build completed successfully!\u0027",
                     "ui:review":  "node scripts/ui-review.mjs",
                     "smoke:test":  "node scripts/smoke-test.js",
-                    "gen:resources": "ts-node --swc scripts/generate-resources.ts"
+                    "gen:resources": "tsx scripts/generate-resources.ts"
                 },
     "dependencies":  {
                          "@firebase/auth":  "^1.11.0",
