--- public/resources/ai-grading-prompts.pdf
@@ -1,105 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R /F4 5 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/BaseFont /Courier /Encoding /WinAnsiEncoding /Name /F4 /Subtype /Type1 /Type /Font
->>
-endobj
-6 0 obj
-<<
-/Contents 11 0 R /MediaBox [ 0 0 612 792 ] /Parent 10 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-7 0 obj
-<<
-/Contents 12 0 R /MediaBox [ 0 0 612 792 ] /Parent 10 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-8 0 obj
-<<
-/PageMode /UseNone /Pages 10 0 R /Type /Catalog
->>
-endobj
-9 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915101052+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915101052+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-10 0 obj
-<<
-/Count 2 /Kids [ 6 0 R 7 0 R ] /Type /Pages
->>
-endobj
-11 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1861
->>
-stream
-GatU3_/e9g'YO;nhQ9O_d3;"^P&F+8EB?ai:9@8%NXcWr,U"K@"otd(?f)#8;R3&OE]VLa#miaGp?h:#6L*SYoRA5Kd@2SU^C)KLBY'ij%UtlUT^1jaQ]lC#E\sYOO^!cW5s1"AE(iP3NT+A;_'GWE,LXaKe-XmX;4j^Gl:)K,Y&4raWpq>OoZ)2>:$M<8GIl0GQ,So([E;-XW(1drO,P[*U?Q<Zr^u4"&<ERQ-`/GX($f389pNjBUhs0*_IsR@[Y4b\,CZ6$O"P2BIK/XQZo-K%97d7m3paVY^K@;$SlNJ=%C\#-9%3HB+VVtOV)qnXme8g*F=r`3#1PpYI8An0Vo+GV7HCo4@]=uAp2E5s13ij<Te5)"1?C,2_eqd\"*lJV^>Hf:n+9@;1(@U>8laSAadUf)&%#ZbN2b=l;>PI]fil$mC^D8s&d@eM*mGSXEBqtfK23ngget`CXJCiU)U,*P/m&M)Pd;c7FAZU^DF^H/Oi`)5h3J$"%\@N.'Ie%NQ>@4VSf]?X[Lr$T3_h%!V#7+@A/.+UeAp:cjK`S$G]\leT@"_bn41"iT=cJq;/7Ts(;;J_5>B>'Vi8!JYYl4,0J;@O_h\BHU<H$BF?eC@!;r:KJ$;gRMB:#U/a\Ehab`DM-tOo>Aic%_O.lOjP!M@B_E]BP5R75WTdbn=TMMJ(6-[[AoShX=!Y.Ag2aXET`&7oc^$:7`pKnCVG,.5m?@*RX$.X7#;=dPjSH,DSb5i%G`1bj`RGDAlM1Xn_F`6"V+7R5,Kung(S?@j&.Ab+WB7:Yi6%!3Xn+0B-V=;pZ'$VpZN'qJf)'bJ:IV_nPj%7UkaMA$j&r^T/1&7T)KDMZNM00%;L0X(nKbb4j$$pOUq/s-*8u=uk`p*os!O0g)2dQXCqjpZ>grsAI/=Ln;?j,>Eh1=pij;FqKm]@sM1(0PEKQV(,+37E,0(e:hmRu`0o9;tDjd^$hYH@7och=od11[p0ph2"6V*Qc3#Mui6B69WJ0MVK:,pB+^CXnl'&%^n4BibMMF=NQe`eHJKZX=E7Kg3/m>q,KZ5E-7R/XL9YVAY!YYM*YO"i$/0=?><&9Gnq&(5DE\i6An1UM*!1?S1,S=jM)<f5A=\8l5TdorcYE>CF7\B-`eu'.#>u3'OXg<B8^5K<cDTC0<]L4reafm0@*]/ubHABH.t4]ZBuF+,Cnd%S?SqREfkeHN?EbnfNaV3lKb_ldOuW"l#\>0&R^_Y8-E\j16coNVkZm/Am7gQV++;"Yc<l)=?e^*m00A-.aeHGjP\`e@;2OVRXo^^j5%;M!.N<I8r2+lFOAjDRU)k'_M[t27+GniA_gP2f%Ut)`Z`mj`u8N<s2'>m\hpG&""h<cndmEEWm*S4%Ki);F.LG5Qc>Y>gi&gg\8>H/)QV*bgLe6#q*WC.L_*7*r]6phc99oSCl\(j4klr4rau[`W:Ppd"i'k9['?fAtF!;G:Ir]KH0$'EMr?@CB<%&hge0&7k,TW=*<ruiqd[KY)URRQZ/9egUo=-XVr9H:pTQ#.*S.Q$j0:@d(]HCA1JHmojYbka_u_Om"!;&VbQIbO30T+/;"L^L2bil4p8TJ25AasqfdD>W/&PlU6YuN4&3C/`foD)M_P-=/_l:iF*_qnVV5nK`4r!N)Q5jYZ4?ZJ#p6YQ?Ba4'`k&jMP'qG]p%J\\O5<hk9>,llBT)I(`f32i=f\<1C3=WT@Pg(.O(%f>'"oo`+EBmm7'WgI4ch>lG!Q(8JiXrrRGAb2PG$qk"HC"eLSFYA;.Pr-R9`DUrBO\-dC;#VO_Tj-S/\O$k$2u,[%4U\G8"8l0C-lZ_T/n#1kP^2[H7PW,P'X[TV$EXWJhqlhg@S]PQ(Ws$,2`~>endstream
-endobj
-12 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1270
->>
-stream
-Gatm:9lJcU'YO<E]Na1.P!t86AD^k6%1K&2>H><E5-2gQEbf]^hHc">IXPjqP_;"G!uQSQ%cl[IR$fnVLaMmbiBOH#Y*X80>=;=^b;D`132,GH\m:o?S4%edHpJR,q=k$M\2jpt9n%[!G7i:3MAH35$FpdQpf*1A_IkoX9XSoBgpo>R;"CZ+;QKDKQ78HdecrU&_*68[d2SMD'ohF394""t8U=+6?tm\CUK,UjrBe\WC&'9M[RN<WE^Fl%1jJM`k#AKp9")iiW6\.QD+I\bddP'P:e[gY)Knf3?QG;Nm/t,HSahG-?2"ldW`1oHjBWtIV/Vreds[.bdB+oi=;'[.n3.oor=bfA6(Q=_:u>TOm_KqlQUqlCBH5(aA/<Sf1:`pnDp\7Y!VmQ`";X)[Qa5VuK@orR*ZX,!>d>C:BG(d%N*jmcoJ-`<i_\lC\?t`.VJIj,?-t$pS!pF[A17oeYqfB%1CLF-#$SO8128@ekfAX=#n0hA,KEgOEW(V%?@u8".c$,?#+-tpLBiWrp4(OnV^qTp+Z]^aXpD6dD8A)03eO0BEn?m/WPEDKdjAbDMsI4%kqe1G:I!XkHuP)EYIuT<=m/W2NOtq"!L^'XdphD1aEd;BJ"_R:MrkQR2)gqAY1;YP$FZprFkDE,mOjOcE>JQo4l-S,S%@!IeL6i/l[@Za3n.4d+&jQ1#N'%Bo8+3/fn6J3>O:r.Lb(LCM7=2(o[)-^/Ze;2"1iB4+*/Y^VU@-t';^C6ZO[iKSg*`5&O7.)l*2"[im(u*T1?M7,00g!QI6Q-ersI#?kS3J%r;UcWbt$<hP]f.$CgZU2^PoQBZP86"dW:14L?2#Y:Q0&kUlE23W6I&2kA9cp1AVQrR#-a@C:2l%sAi/bP$m2Tj-iZ:X;YB4WXV)C:PnA_VXk/#Hd_t(hJ='Yt\M>"!mRO9gY+3=$.nDA9N5tGY&nR']IjKs$07*\YZA4oep+"hW9>":13p!K%tT-h>9V`TaD8I$q423B2XDHLGM#hO(BiM?<nMl/Ub8Njh5Of=sIK&?nNsad2_de$i#t/a)oYIC*cLJQ+P0D8paiUllZHNBM8)'$A\2Q)l.%+T,sh?HLZVdijap&^7mJLe2>6M95M^Xn`3tNr_GACGNM%-/Z8[+Q>.Bt/b:cLHb\/4DQr^(>p^RAp[9dq1^qY___2HI3'35-j)_GRquca(eGifFW#R0!ppJP]:jogSr]S&2`Il.)l+;lR`E@h7at]t/OKTRt]VBf3iS=t:ntn+3r!1ZGpl5~>endstream
-endobj
-xref
-0 13
-0000000000 65535 f 
-0000000073 00000 n 
-0000000134 00000 n 
-0000000241 00000 n 
-0000000353 00000 n 
-0000000472 00000 n 
-0000000577 00000 n 
-0000000772 00000 n 
-0000000967 00000 n 
-0000001036 00000 n 
-0000001319 00000 n 
-0000001385 00000 n 
-0000003338 00000 n 
-trailer
-<<
-/ID 
-[<387afed868c59c1999d1a7f0ba07e5da><387afed868c59c1999d1a7f0ba07e5da>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 9 0 R
-/Root 8 0 R
-/Size 13
->>
-startxref
-4700
-%%EOF
