--- public/resources/ai-parent-comms.pdf
@@ -1,212 +1,0 @@
-%PDF-1.4
-1 0 obj
-<< /Type /Catalog /Pages 2 0 R >>
-endobj
-2 0 obj
-<< /Type /Pages /Kids [3 0 R 4 0 R 5 0 R] /Count 3 >>
-endobj
-3 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 6 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-4 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 10 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-5 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 11 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-6 0 obj
-<< /Length 850 >>
-stream
-BT
-/F1 32 Tf
-0.424 0.388 1 rg
-50 720 Td
-(10 AI Prompts for Parent Communication) Tj
-0 0 0 rg
-/F2 18 Tf
-50 680 Td
-(Professional, empathetic communication made simple) Tj
-0.424 0.388 1 rg
-/F1 14 Tf
-50 640 Td
-(Transform difficult conversations into positive partnerships) Tj
-0 0 0 rg
-/F2 12 Tf
-50 600 Td
-(This resource provides prompts for:) Tj
-0 -20 Td
-(• Positive progress updates) Tj
-0 -20 Td
-(• Addressing behavioral concerns diplomatically) Tj
-0 -20 Td
-(• Celebrating student achievements) Tj
-0 -20 Td
-(• Requesting parent support at home) Tj
-0 -20 Td
-(• Scheduling conferences and meetings) Tj
-0 -20 Td
-(• Following up after parent meetings) Tj
-1 0.8 0.94 rg
-50 400 560 80 re f
-0 0 0 rg
-/F1 12 Tf
-60 460 Td
-(💡 Remember: AI helps with structure and tone, but always add your personal) Tj
-0 -15 Td
-(observations and specific examples about the student to make it authentic!) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 1) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-7 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica-Bold >>
-endobj
-8 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
-endobj
-10 0 obj
-<< /Length 950 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(Positive Communication Prompts) Tj
-0 0 0 rg
-/F1 16 Tf
-50 720 Td
-(Celebrating Student Success) Tj
-/F2 12 Tf
-50 695 Td
-("Draft a positive email to parents about [Student Name]'s recent improvement in) Tj
-0 -15 Td
-([subject area]. Include specific examples and suggest ways they can celebrate) Tj
-0 -15 Td
-(this progress at home.") Tj
-0 -30 Td
-/F1 16 Tf
-(Academic Progress Updates) Tj
-/F2 12 Tf
-0 -20 Td
-("Write a supportive email to parents explaining [Student Name]'s current) Tj
-0 -15 Td
-(performance in [subject]. Focus on growth, efforts made, and next steps for) Tj
-0 -15 Td
-(continued improvement. Keep the tone encouraging and solution-focused.") Tj
-0 -30 Td
-/F1 16 Tf
-(Requesting Home Support) Tj
-/F2 12 Tf
-0 -20 Td
-("Create a friendly email asking parents to support [specific skill/behavior] at home.) Tj
-0 -15 Td
-(Include 3-4 practical suggestions that fit into busy family routines.") Tj
-0 -30 Td
-/F1 16 Tf
-(Thank You Messages) Tj
-/F2 12 Tf
-0 -20 Td
-("Write a heartfelt thank you email to parents for [specific support/contribution].) Tj
-0 -15 Td
-(Mention how their involvement positively impacts their child and the classroom.") Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 2) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-11 0 obj
-<< /Length 1000 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(Diplomatic & Professional Prompts) Tj
-0 0 0 rg
-/F1 16 Tf
-50 720 Td
-(Addressing Behavioral Concerns) Tj
-/F2 12 Tf
-50 695 Td
-("Help me write a diplomatic email to parents about [Student Name]'s classroom) Tj
-0 -15 Td
-(behavior. Focus on partnership, specific observations, and collaborative solutions.) Tj
-0 -15 Td
-(Maintain a caring tone while being clear about the issues.") Tj
-0 -30 Td
-/F1 16 Tf
-(Academic Concerns) Tj
-/F2 12 Tf
-0 -20 Td
-("Draft a supportive email to parents about [Student Name]'s academic challenges) Tj
-0 -15 Td
-(in [subject]. Emphasize effort recognition, specific support strategies, and) Tj
-0 -15 Td
-(request a meeting to discuss next steps.") Tj
-0 -30 Td
-/F1 16 Tf
-(Conference Scheduling) Tj
-/F2 12 Tf
-0 -20 Td
-("Create a professional email inviting parents to a conference about [Student Name].) Tj
-0 -15 Td
-(Include the purpose, available time slots, and what to expect during the meeting.") Tj
-0 -30 Td
-/F1 16 Tf
-(Follow-up After Meetings) Tj
-/F2 12 Tf
-0 -20 Td
-("Write a follow-up email summarizing our parent conference discussion about) Tj
-0 -15 Td
-([Student Name]. Include agreed-upon action steps and timeline for check-ins.") Tj
-1 0.8 0.94 rg
-50 180 560 80 re f
-0 0 0 rg
-/F1 12 Tf
-60 240 Td
-(🌟 Pro Tip: Always personalize AI-generated emails with specific student examples) Tj
-0 -15 Td
-(and your own voice. Parents appreciate authentic, individual attention!) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 3) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-xref
-0 12
-0000000000 65535 f 
-0000000009 00000 n 
-0000000058 00000 n 
-0000000127 00000 n 
-0000000260 00000 n 
-0000000393 00000 n 
-0000000526 00000 n 
-0000001428 00000 n 
-0000001487 00000 n 
-0000001543 00000 n 
-0000001583 00000 n 
-0000002585 00000 n 
-trailer
-<< /Size 12 /Root 1 0 R >>
-startxref
-3637
-%%EOF
