--- public/resources/assessment-rubric-template.pdf
@@ -1,99 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 10 0 R /MediaBox [ 0 0 612 792 ] /Parent 9 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/Contents 11 0 R /MediaBox [ 0 0 612 792 ] /Parent 9 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-7 0 obj
-<<
-/PageMode /UseNone /Pages 9 0 R /Type /Catalog
->>
-endobj
-8 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915101052+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915101052+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-9 0 obj
-<<
-/Count 2 /Kids [ 5 0 R 6 0 R ] /Type /Pages
->>
-endobj
-10 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1467
->>
-stream
-Gatn&9lo&I&;KZQ'mo2pl,9k5;6m8ra0Cgbdrt$+8pY/sJ-^s!M@TmkV5V/C$]q`,<l<"[7l0jpKReiin*[n`@#U^'rll.mX!O:N8K]1\>XO[7;HWo"YEpVD:6K5N`>iKV";g++;0nWWaW*d'>bogO7[NB$R"3FWE1c3?lSI<jiP#Fc)YPI&loL\ra;jY$)4Y"na.6.6Hhb9X=cpO`X&;jAh$[>6L?/qb[^R</hW*cR<'%4V`pm6!Tg>($5?^<3]sVI3#7I<+Cef/X2ASJS.no'_Y$-89R<LnR"'ck]Zq(EV8;p:+l?&U].:V<q@(8&?c6M]u`=>'khLbaZa"crl7sKJ_Vg9=YLab?$6Vpsm&WB((3OQZ;QMB+^A!l%Ei7_0HDrDMkHq7CX.2+al)1tn-=b?5[_`J>Xc+%m[h"XcE.K$?&5)iXCASqXoK($M&%0nO#LM\1[#GE>qc@J0XK4\7tBsnuR55Sl6+:%/mr#'O:T%t<9+E+gQ6)Pa6p0jd6l#@'dccEL/*Y1"EOp1c/D01a3ce<1]B0$!mpV?#J-..i?$>8l)CO@,6>JL\LQtUD*.A@A/?M_6:*\n[4Zk^:T`]_;b0jI6=pUmXI'T,BU"S)-$VR*2<:(T>5YfHs%AH@"/ffF(A$?c^%)-#31\B-J)lbF*c]AX\]9o0!)gkn/Bb5jRO'W1^MGeH4m4ImicSl\#j>4:DuTnPOt:97KWFm'9JEu\185#?Heq2p[4Yj=rG]i(+2=_pCe\!B,"G5Tdhs2i61.!>03*BISAH9'5[O5"\(c]$,J(B7dN@\c=r8/Vc@_Y.fs+K<X:ReO7Ca_^nT"WuJ1Q#bK`lCdAO>]'ss'%^!5oNPM>9%t-22cpZF16+0+o+j#,0G:is!OESZ18%-ZAH+sI83#[uE1L*r"tq&m6dc'W-hH#4N-a!'-q#m],1h1]W"4^.#3c#FaXVN6NcF@Bdj^]6pZ]WuIPTY+FC5:lV(Zd72aET^/AM%Ma(/h*i9H[Z.Pusc<Ne>)dfIVkD-o.""Vm(QEFMq98"\3T<gk#_LP3&lgO?TQ!;1gu,3.j1IY[=]&R*pIU<72,L8M=%PX#It'/o(;+p4X%G%$$O$.J8r=)F^oK;+Psqkd>+Ej!>OhA\JoZN;BN%A-4T6Ud)A4@^"=KdN/Veor35!QfFK"iOo`^n(nJEaKZlJ2]rl:dSb\b.%o\1<r]lY),kTB2DU??CZ+fcu5k\E/ak*03Q(jj*;T]'#lVo^ef^:dETgR];`L>\\cKV`m?>a#B`?;M#7Tch[V7<!oDO(4rfX_.?g$T$=fB>26<oW^r+n/VQL^Z)5-3lfcPr)^8A:\HPRboC*O^=!Yqe$(?7a`SiqjXhc`sD:Jbsj5eKB2FCf8%Oo'Uk3E7J+[6dq">@[]Q:?YE4]&gAEW!uSi4#Dqan.>!VJmu^!WbGY11p;C2A9G1;c&)_nqjO[?"=O3NeX(7g~>endstream
-endobj
-11 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 527
->>
-stream
-Gat=gd;GI.&4PLN/+-md.25^((\3k_M,^HP;jDf^0`sBip;T"[jR517X)>N[8/k\S1lQ(-k2$\@q2"jK#2lZ$QNmPRTkYbA,Y.MJpXSP?-'"DJ!o#QlA<*L-b]bSO-N7'*]>L=73`jG#)P8(XPm3<+25?$fpou2"NPWddNK3IH/RhLPDV,hae6qIq'<1aEeJ"s`M;]hSs#;oJH8u3D-jno;Po^3`['HHXCkT0:e(W/f*LteugG+k8A(S>4R=cUD3032l\g/qM+UTJt[a!D*8XYo-_/nE(J=D*g-er1NrAMo1ifd5$bL!l*La"aDO$k8geOr^-A*pZ2Gh'&hLh">N");Qks-+\jc47lkVsPZuRU=@&-28KU'\jFS9RG$T=g=/]6lmpW;bQer"L0RDi7"mOACA//#[p4q[p"@>j<VVDM59F)_=KUt7hH0saXhHH+7O!9i78b>HI<rtf&2rTL;N#*!*QW_TtHEH[9P!0'UKK7s-3[Ug2>tWJ[TfIj%4f9B4k;9!FYR*4_,c\~>endstream
-endobj
-xref
-0 12
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000656 00000 n 
-0000000850 00000 n 
-0000000918 00000 n 
-0000001201 00000 n 
-0000001266 00000 n 
-0000002825 00000 n 
-trailer
-<<
-/ID 
-[<a4b4dc2448b17b85ccc6e82e1e91d200><a4b4dc2448b17b85ccc6e82e1e91d200>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 8 0 R
-/Root 7 0 R
-/Size 12
->>
-startxref
-3443
-%%EOF
