--- public/resources/assessment/assessment-rubrics-pack-v1.0-en.pdf
@@ -1,80 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 9 0 R /MediaBox [ 0 0 612 792 ] /Parent 8 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/PageMode /UseNone /Pages 8 0 R /Type /Catalog
->>
-endobj
-7 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-8 0 obj
-<<
-/Count 1 /Kids [ 5 0 R ] /Type /Pages
->>
-endobj
-9 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 901
->>
-stream
-Gatm9gMWKG&:O:SN;=)GPtee->X(62,=P5TUdaO+?T-)8Dp$&P?/i<8a*1Y$\KRD6&JB]O@EpkhB*SZX\f>eS3P\2pT?\Eq>_l<tI*7<?!`njTDWJ0e@0EpI9>RuIftQWo+]ro$#>kViis,K&jCUf*S3uQjA]lO+$a*k"Fko+EGc;B,Q"-NH>BM.oLJ,X:*^b;_n])R?MQZ-X6.>nhB'gm>$f`8k)jjVL1qUj@U]F>SQ4imhBMiRmY`l#A!Rmbo<;g"W,1lJ*M?J0lAaH:YB%Ir3k1B9Q[HdR9X1$iMc];[U(kC_ZTcZ7%c/#p+ggB?P]CItEd1mO)58jj4*oeo6Ei7O06O\UgIREM$osV\M>4_fmKS!)UG^pOF?rB?U^_]Dp4=tbV1WHodD32R"/^8W2?ojiFc[?AHpD]o@!4<j<2X_kc8#D#%:QR"bK%7P*Yi&&SkCjP<X\Po*K\>/h]d7#"pS-TRA:0>!c&g&F@1kB/B4UJOTT8a@H^S@W]7<T9<90i&XB1rSfZ6No:EdKq%4f@=j(2\H;/QML)5LEG?N?rPfj<GYBZWfO3=^$'#a<66NV:ju^^3q]E0Z_(f9O:n6/^5]:>hctIm3aZ>uTKu*/O@/Wmhb-eM9A,4eS%9U]P&2AT_:(`lQH?C]B9DVEhV_,Di4.Emm>N<oRCblg"Tb5meL]ONK?I7+_^Qqt;,cqSn.0bhch@>Ws$6Dk.*H2sWIJmdLNc4(/EHV4^F%Fg[aT[87V686*9-.J3h0<?dB[Z.Q8if(4:k0fbmJ[o\=6ph4@nL?\B]bF?7>]p@i#8T!8I+U274kf3U9AIA(GpeY_P>G9cC%O@46)FU8l@,Oan/mnn.DlK3S>UY7.'piC-a`X4amFoC36'@4Y?ELZTK@+$u1CY]~>endstream
-endobj
-xref
-0 10
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000655 00000 n 
-0000000723 00000 n 
-0000001006 00000 n 
-0000001065 00000 n 
-trailer
-<<
-/ID 
-[<7dbace2713b1c22e2ac80b1f4293dee2><7dbace2713b1c22e2ac80b1f4293dee2>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 7 0 R
-/Root 6 0 R
-/Size 10
->>
-startxref
-2056
-%%EOF
