--- public/resources/assessment/grading-prompts-toolkit-v1.0-en.pdf
@@ -1,80 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 9 0 R /MediaBox [ 0 0 612 792 ] /Parent 8 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/PageMode /UseNone /Pages 8 0 R /Type /Catalog
->>
-endobj
-7 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-8 0 obj
-<<
-/Count 1 /Kids [ 5 0 R ] /Type /Pages
->>
-endobj
-9 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 569
->>
-stream
-Gatm7Z#8;M'\rX%LJ1:?W$G)?ElSr4QOYI;CgLeCO]Yh`9%%^1+TL:GfhTp_:kPJb`8U?-GOH%&@!)0_="Wp8`4O>8>bJ_3Q`3CK'q`Y%Nu:/VjolMn6,SAP9[<Gr8'5k<'a7Jf1:`(AB>h][W3;&[-!gTqHX7kAf#o^QD2bT:j'CEYJC;!af#0)'hbh,P<PQrW;`Q@gb&"kb$_;D%:]:Ti1OjM&HZ)c,'=Bb+cD*+:Z_NS1)bU:(f9HY'_:I6Jm^1iP<OGOK4ion!([,+aTo>P*PJolL),_VuR;o#qZ'=>Gq*s_@*&!&+MIE9@c>.MOdMb#;H-6QgMS,m+2R"9M=n+Vp`GOhNM2_i2qmA?t]K>'1Ti3-"a2\=<5";9F>ha!d!Nn%Bo\.m$_-:%?1P-sceQ%3[Xb9g;$jYrS+ha8M]Qd/iUR'*;h:Uqa?hC@+Q1lbsH]<Y$LX'X&j_Gm1W+f7#j6Z[#qu+d8(hkoB`W'V\1p+J%l<aAPDV;c/FlS&kP&Qu#UlnPYqrlQ??/nrBHZIDd9WS>>-H+"bOKl&QA2/u(M^+,404fq]%f~>endstream
-endobj
-xref
-0 10
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000655 00000 n 
-0000000723 00000 n 
-0000001006 00000 n 
-0000001065 00000 n 
-trailer
-<<
-/ID 
-[<5881f3f4124ef4eddda3bd99606a6235><5881f3f4124ef4eddda3bd99606a6235>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 7 0 R
-/Root 6 0 R
-/Size 10
->>
-startxref
-1724
-%%EOF
