--- public/resources/assessment/report-comment-bank-v1.0-en.pdf
@@ -1,92 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R /F4 5 0 R /F5 6 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/BaseFont /Helvetica-Oblique /Encoding /WinAnsiEncoding /Name /F4 /Subtype /Type1 /Type /Font
->>
-endobj
-6 0 obj
-<<
-/BaseFont /Symbol /Name /F5 /Subtype /Type1 /Type /Font
->>
-endobj
-7 0 obj
-<<
-/Contents 11 0 R /MediaBox [ 0 0 612 792 ] /Parent 10 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-8 0 obj
-<<
-/PageMode /UseNone /Pages 10 0 R /Type /Catalog
->>
-endobj
-9 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-10 0 obj
-<<
-/Count 1 /Kids [ 7 0 R ] /Type /Pages
->>
-endobj
-11 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1587
->>
-stream
-Gatm;9lo#Z&A@sBFSS>;0p.-B"?G.l\(4Ou=#MD5(G:?iJhb\ER2LS/?b\aV"Dc?k4/)"]g#Jbj?GCsFj+9X*Y5K7$!4NX:n@K(2#Eoj,^[K)$:Ulm!r4$_3RpcH;O[kc&XRGodG7[*u$(W2rT^R2X56<j[+?T#:oJKdpafYl?DsVeAZ.f)0AK`pY.e4R5L0H!jmpP\hP:P\`0D^"Z22J\I*IDtnV"QOq`GU+ZXa#FA7edF;?F!\^0*7b'.#fOk>Ls3o)c5)O$0hI`,`pslZ[QO@^;9.!FPsX!3_g-N*t>n;JA13#Ic*PBjGeNopFm>[J]j$h.ZmHe;9LH?Q,Ki8JuJBcN/_UM\(,V[f`jnIM8bAW$V61,/(P'/(E8%@h7L;5^OE*l)\]o)F3WTJWrYGt$.IaW_)/5*'O+Aj2'37l$eslk;_Jt/I'in!Of*P1ci<USWPf1eiuFX3)T!ITl_I&_%U7!cKPeX6I/cn]jm:(Vs-N(]SoM1%%*BO'U59?[RCZf4=''Ul0k/#=I!g(el,30iER6r4i:s76T*h8O_(KhAg9WVNTn)@G'WeiOJ>TnW/-Dt6f3Jlq%=!@:]fjsDj&eNU:=3gu1b)<j\Y-``9Re?6-nO+L,$%iI"[<Ts5AJXY5P!YPkJM;0GL%R-DB%Cu\70JG'a7#R<[UnB15_;]!p8Tr;=NBbLJ_9s_k3j\(IW\?BRXnAU4$+-m>n8$[F7Q3296XO>Inri+1)YrE#rJ:(C<^1M)u'1\,j:Y*7`(mX6;W"j/OL6YjOb\Vj$S?LKQQ-!].qQ"ahpY$;Ya.#_Qbp#[o&qe#)8daa%H"fH%3Nat)(Z3?:N*1FVH'VE<m\&9MsH=F$Hd_G+0FSG8CORHA'm)is&hDU+2;=:dIn.#EG-0Kg=eaCMaT\6iX`0Lp\WAc/1V<m]#e08d[l?\t$/4QB+%.>q>`GrLun"7PF%g\37H=ql\T.83Y]F!ES;:f_[Y/r[/.=\/30Rs9MS'%VbARZ._I6??aTmW.'6_'ftJl9m7"kIT<:$WT\4Zk_-e4AajuJM8bM<['6(]4+^HQUN>L/LI:b:^F_BMe7UUc$u))MCUKsnfkp:<&/<\pVLujrm=[nmbB9Ts'2r2N/Gr<eLBOAk-rcC[QdbkldTmG"qilfXbed@HYFem/;9;r@n8n^9LG$WA$5OL=-dW4-`/mCZWWR@FR-lqo?K_uY9\?<`F<C<40PL,Ed_9aKFtjk3Z/`,8=4"^D44KJWRLGcWQQ1!XAX;R:J4XSp79X&Mqb[Y\g-AH*4a#l>OeGN+9"1Cq$3C/Xh'+(_bOT$L]9;Dai<7JoM-6J+&V@Jc=e7PZd&C',=N@_Q!dB'gIa^o8XnbK#\JoBO[a"3G-:iQ@^/o<M'QkK<h"VX8Y8n4#ZB(ZP)6d!XhN2?^3sQr4U9CpKXfb!5h_B-dl,<*Kl/ohp/I\T/BnMo0OCJ9O]%ZZ\;`!;g0=43*c@3FcJe(i'Y>QG?Lr&QKD'=`8O4:+XZ3V\h'/_BY:Xig=,@eF]9RsUilck0?oPRN]7/L4p?DDp6>,S/6^R_tKLsZ1VOeH+G""T2gHUEMY&NnaFhD@G~>endstream
-endobj
-xref
-0 12
-0000000000 65535 f 
-0000000073 00000 n 
-0000000144 00000 n 
-0000000251 00000 n 
-0000000363 00000 n 
-0000000482 00000 n 
-0000000597 00000 n 
-0000000674 00000 n 
-0000000869 00000 n 
-0000000938 00000 n 
-0000001221 00000 n 
-0000001281 00000 n 
-trailer
-<<
-/ID 
-[<393c62bf45d0739b2f876658aa75ca8d><393c62bf45d0739b2f876658aa75ca8d>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 9 0 R
-/Root 8 0 R
-/Size 12
->>
-startxref
-2960
-%%EOF
