--- public/resources/best-practices-guide.pdf
@@ -1,74 +1,0 @@
-%PDF-1.4
-1 0 obj
-<<
-/Type /Catalog
-/Pages 2 0 R
->>
-endobj
-
-2 0 obj
-<<
-/Type /Pages
-/Kids [3 0 R]
-/Count 1
->>
-endobj
-
-3 0 obj
-<<
-/Type /Page
-/Parent 2 0 R
-/MediaBox [0 0 612 792]
-/Contents 4 0 R
-/Resources <<
-/Font <<
-/F1 5 0 R
->>
->>
->>
-endobj
-
-4 0 obj
-<<
-/Length 220
->>
-stream
-BT
-/F1 24 Tf
-50 750 Td
-(AI Best Practices Guide) Tj
-0 -50 Td
-/F1 12 Tf
-(Research-backed strategies for implementing AI in educational settings) Tj
-0 -30 Td
-(This is a sample PDF resource from Zaza Technologies.) Tj
-0 -20 Td
-(Contains guidelines for ethical and effective AI use in teaching.) Tj
-ET
-endstream
-endobj
-
-5 0 obj
-<<
-/Type /Font
-/Subtype /Type1
-/BaseFont /Helvetica
->>
-endobj
-
-xref
-0 6
-0000000000 65535 f 
-0000000010 00000 n 
-0000000079 00000 n 
-0000000173 00000 n 
-0000000301 00000 n 
-0000000586 00000 n 
-trailer
-<<
-/Size 6
-/Root 1 0 R
->>
-startxref
-645
-%%EOF
