--- public/resources/classroom-routines.pdf
@@ -1,206 +1,0 @@
-%PDF-1.4
-1 0 obj
-<< /Type /Catalog /Pages 2 0 R >>
-endobj
-2 0 obj
-<< /Type /Pages /Kids [3 0 R 4 0 R] /Count 2 >>
-endobj
-3 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 6 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-4 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 10 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-6 0 obj
-<< /Length 850 >>
-stream
-BT
-/F1 32 Tf
-0.424 0.388 1 rg
-50 720 Td
-(10 Classroom Routines That Save Time) Tj
-0 0 0 rg
-/F2 18 Tf
-50 680 Td
-(Proven routines to streamline your classroom management) Tj
-0.424 0.388 1 rg
-/F1 14 Tf
-50 640 Td
-(Transform chaos into calm with these essential routines) Tj
-0 0 0 rg
-/F2 12 Tf
-50 600 Td
-(Research shows that effective routines can save teachers up to 30 minutes per day) Tj
-0 -15 Td
-(while reducing student anxiety and increasing engagement. Here are 10 proven) Tj
-0 -15 Td
-(routines that every teacher should implement:) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Morning Startup Routine) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Post daily agenda and have students complete a warm-up activity while you) Tj
-0 -15 Td
-(handle attendance and individual check-ins.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Attention Signal) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Use a consistent signal (hand raised, bell, or call-response) that students) Tj
-0 -15 Td
-(respond to immediately. Practice until it becomes automatic.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Material Distribution) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Train students to collect and return materials efficiently using numbered) Tj
-0 -15 Td
-(bins, designated helpers, or strategic placement around the room.) Tj
-1 0.8 0.94 rg
-50 200 560 80 re f
-0 0 0 rg
-/F1 12 Tf
-60 260 Td
-(⏰ Time-Saving Tip: Invest 2 weeks in establishing routines thoroughly.) Tj
-0 -15 Td
-(The upfront time investment pays dividends all year long!) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 1) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-7 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica-Bold >>
-endobj
-8 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
-endobj
-10 0 obj
-<< /Length 1200 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(7 More Essential Routines) Tj
-0 0 0 rg
-/F1 16 Tf
-50 720 Td
-(4. Transition Procedures) Tj
-/F2 12 Tf
-0 -20 Td
-(Teach specific steps for moving between activities: clean up, push in chairs,) Tj
-0 -15 Td
-(walk quietly, find new spot, wait for instructions.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(5. Independent Work Protocol) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Students know: Try 3 things before asking for help, use whisper voices for) Tj
-0 -15 Td
-(peer help, and what to do when finished early.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(6. Bathroom and Water Breaks) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Use hand signals or passes to minimize disruptions. Establish times when) Tj
-0 -15 Td
-(breaks are not allowed (during instruction or group work).) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(7. Group Work Formation) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Practice moving into groups quickly and quietly. Use numbered groups,) Tj
-0 -15 Td
-(assigned spots, or random selection methods consistently.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(8. End-of-Day Cleanup) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Every student has a job. Play music during cleanup time and finish) Tj
-0 -15 Td
-(when the song ends. Check work areas before dismissal.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(9. Assignment Turn-in) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Use consistent location and format. Train students to check for name,) Tj
-0 -15 Td
-(date, and completeness before submitting work.) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(10. Emergency Procedures) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Practice fire drills, lockdowns, and other safety procedures regularly.) Tj
-0 -15 Td
-(Students should respond automatically without teacher direction.) Tj
-1 0.8 0.94 rg
-50 80 560 60 re f
-0 0 0 rg
-/F1 12 Tf
-60 120 Td
-(🌟 Remember: Consistency is key! Use the same routines every day until) Tj
-0 -15 Td
-(they become second nature for both you and your students.) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 2) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-xref
-0 11
-0000000000 65535 f 
-0000000009 00000 n 
-0000000058 00000 n 
-0000000120 00000 n 
-0000000253 00000 n 
-0000000386 00000 n 
-0000000519 00000 n 
-0000001421 00000 n 
-0000001480 00000 n 
-0000001536 00000 n 
-0000001576 00000 n 
-trailer
-<< /Size 11 /Root 1 0 R >>
-startxref
-2828
-%%EOF
