--- public/resources/communication/parent-email-playbook-v1.0-en.pdf
@@ -1,86 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R /F4 5 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/BaseFont /Helvetica-Oblique /Encoding /WinAnsiEncoding /Name /F4 /Subtype /Type1 /Type /Font
->>
-endobj
-6 0 obj
-<<
-/Contents 10 0 R /MediaBox [ 0 0 612 792 ] /Parent 9 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-7 0 obj
-<<
-/PageMode /UseNone /Pages 9 0 R /Type /Catalog
->>
-endobj
-8 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-9 0 obj
-<<
-/Count 1 /Kids [ 6 0 R ] /Type /Pages
->>
-endobj
-10 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1732
->>
-stream
-Gatm;d8J^Y'Rf-pmd`\PK!??:$/L@Wb'(8<G%TDN1OUB..(4^r*6FIWdL=TSh]pfG.<qCAgW9j,@I%6jHd90'!VuKnmUrgNH97h,56;0YR`m4:]SImVkhY=Xdm;>C3>thD*o^V0H%%RU"s`^:OZshHaU>soFAD]3!;coRX\Z4V')RCt7X!ZC[(H%7RHDrE]-S+N?0?#e;eg;GP6[PHfU2]so5382WoNqnm:DY-Y/:[Vl+8VkR9oTdct^kL;CKY-i\k==F/:<_bGCA$-CIalQ,6MEi1Y00L830ND7D4qPiSjO?T-U1QqYV8OY#'o<$@FQlV'F"D;qa"rKc!m3,ZNk`c3sVP(u`*@6i[pGTPgS09>L=^^\<Oe73^fpM,E+W%'fq9U&M\U_DKkgk>W<DcaHaQ-7Rl[h1O^;_RY(;AEtVjL05'\4:iSFNsC=c!I#ERSm;X7*G[8+1ZA_(45-<HC#QDQ^cnZAn]:[P4?ugJ:gJN=b'p#[R!(KJ,INa7t6%J-'oa4GiVTtqrEc2.q(4&X'd8Ys%3[E$^]gW,umMZRf2:.0dZj9n2`4/T!\Ag;&cKh;(JVtk_Bl<Dj;#";pnC8LT[_Qo=.GgN>$,'=Mrp`#,;d$SG^JTf8sWFUs/<e-RYI9`gp_g"H'o&oNom]D*)BspndHUF@@],%8X*0apa3diMR%*DUKXc;jF'2)cNSXn*[-]??35X-6skfX4[SNbZug,($CM#W^YsVqF'juDq(pi6U[nsK/WY!lKe^lM[!^Z>/Ypr1s+pbe4'PL)jLLCQ*ReTe0=&a6;MjOBA-g!@j^7ih+>)42i@OH+hL]/iEF[:=&qb@PQB\6GSbNh-[&N8D;aqH=0G2>Zm6c[A4#.:V&.3S5[<5sNS[i=eh*K"mWQT+p*`n:cW3XkD!(kk[TVF')2E1"j4T]?a;$qFp7ql5>q2:!4Uq?upCUVlp=kO<0_U!JV-MT2[4A4B*:@%p8]KhcmdgG.2LB.kX=PE8h7_*k%u,F0bSE0J7:F4=oXosZ1sYrUSB5D95lGB<Bd(m+l^FNQ-/hLoQ@)2GK=sG."#h?Z;,re,biZb0)8$B2H[;uhpZ7<->OtA$=1aW/A'D`LIf_PUcfZgZe]aHis3*%_bNX"J?K$-Y6i?ufcE:]Q0CES\_]-J&HSTI4L6)7'2ktrq!D+29+=&U'qAXOaA9gPYoGA5S:DUJ@M,5P2%&<#bM#RVWkn7cLpc,C54q4`0XL5i9ghWfSIeAIh<TU3$Q]1)28h;7;O`3h9Y>H2>rmlM[[S@T-qZHkA@rTL>'cL?mk-bZnE.?pVbn))(fH7ti;^YTO;p3'kQ+/ltF5-8P"`U0(KJPa)+>G7cGmZX<f)\PJgsQjQcd[/p/a>dRU@Yc_(NPLMg$X&;h"]U%*d3N&hZ\Dn]cm%B#QA<sb[tX$-$P=\620a=+.ustEtqr)/o4T&R.]7]K!X2b+p!qS7uK7LON^rHRKJ&4Y;s[qK%\CFk.KZu,Lk1oQ_npWgVT!F;GebZ,0ndUgA7,e3M-JWo3!c)g/^-WaT&lP?!Pb\DPiNO[:>Wt\cgWbZ\%.6d-Ym?GL\X;BIiu=n!_KQV,siRa@1M7GPC/&l_c*82Hk_P:`+%?m;?9+!R-k\_+iZ7bua3KU$ot`p>TUB`imfPloNeUf8.<`X&.S$U#j)aD9r97J](HBq(I(<PACQefo0dS_s549&h,stIBV^l[+a.eC0QW9`mDQ:'Wul!~>endstream
-endobj
-xref
-0 11
-0000000000 65535 f 
-0000000073 00000 n 
-0000000134 00000 n 
-0000000241 00000 n 
-0000000353 00000 n 
-0000000472 00000 n 
-0000000587 00000 n 
-0000000781 00000 n 
-0000000849 00000 n 
-0000001132 00000 n 
-0000001191 00000 n 
-trailer
-<<
-/ID 
-[<5903eb042af08e2efe91ea6cce449436><5903eb042af08e2efe91ea6cce449436>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 8 0 R
-/Root 7 0 R
-/Size 11
->>
-startxref
-3015
-%%EOF
