--- public/resources/communication/translation-helpers-pack-v1.0-en.pdf
@@ -1,80 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 9 0 R /MediaBox [ 0 0 612 792 ] /Parent 8 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/PageMode /UseNone /Pages 8 0 R /Type /Catalog
->>
-endobj
-7 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-8 0 obj
-<<
-/Count 1 /Kids [ 5 0 R ] /Type /Pages
->>
-endobj
-9 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 605
->>
-stream
-Gatm892F=s&AIV:pr'"%V%']]H)u=DWedZG"XuqomUhW)S(O_-"X1g6YERcb9"+h$Q),Hip@k#D0Z;qSiO`uYZN;N;%#gs[^dKGE>iJ[:q`/0`Rotbn`P^e6Mdir%b7%7<`CZXG@QY5QfQn<PfG!m$TX=feK8Ue;o,`$\=K,RN]eUKBe9#6W0oGEJhY2u\iV^uij?(q;W4aJ0F"Erj%icZcgBMo3hAF_7gK<^]6"tGhP4jgV.ahmLO%/D\.r3"mnf.@:dQ20tcF[afS*8o0UUh.pWc>[I:ZNtN&k&7&,%6:.HmOC@Qd4SF4dHW(o&n:f9/lIFOhoZ5j5tSYDj&.Za>1t*4>?4"Fi1Kn7i/&!4(Eb*3L!`7cu<l&.IX%,8]LBSh)i]&8cG8J9B>%@9'+W.kPDX:QCN;U2G*q6p8XgJ/=`^tUb6)tDlncO&1V2Ef6@f87)ICPdJD><\/L*B8qG(&Pk0>grb?l:9^2_0GSk)YnIa40PV:<a84/\8f4aB6RL!8V3b[304"8'/pAsj9W"Q+0VKEKb<Ya:!GF4p"fdM5c\ph]:7TPWJ=iOW[Hm<J#3cr/4k:G%K.jfGSMX)-l^AoS`2;%~>endstream
-endobj
-xref
-0 10
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000655 00000 n 
-0000000723 00000 n 
-0000001006 00000 n 
-0000001065 00000 n 
-trailer
-<<
-/ID 
-[<b670962000f071593c435667682ca954><b670962000f071593c435667682ca954>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 7 0 R
-/Root 6 0 R
-/Size 10
->>
-startxref
-1760
-%%EOF
