--- public/resources/formative-assessment-checklist.pdf
@@ -1,86 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R /F4 5 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/BaseFont /ZapfDingbats /Name /F4 /Subtype /Type1 /Type /Font
->>
-endobj
-6 0 obj
-<<
-/Contents 10 0 R /MediaBox [ 0 0 612 792 ] /Parent 9 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-7 0 obj
-<<
-/PageMode /UseNone /Pages 9 0 R /Type /Catalog
->>
-endobj
-8 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915101052+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915101052+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-9 0 obj
-<<
-/Count 1 /Kids [ 6 0 R ] /Type /Pages
->>
-endobj
-10 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1379
->>
-stream
-Gau0D9lo#Z&A@sBoPM^9_3,?H\2jOZ\^oGqK\B[Y)5NQ81R_0Q7M,L`[mtfKZ]lS:R=]IEKfj[1rSHjX63uciej#?"i;Ee]*^U,WQt-cJSJW\LE'Lo.IM$Jc]H0=f(t$TMXV26*0H<>*MQYuG#b>Bd$mO'o"e?=pNc*<["e?Su'YAF]aIV$*G)"g:I6:)poq6EC/70OT(k]gB-nX6"0UW;dloUTI^uS:)n,EijqFn*%aPt#S=\rHe"L>kkh>Mg58*b68P<"$,$oQ01+]B<IZOE$b4C>4&(3>LJ$g9!nW+P5NC2X[`&nf%H1:l"%*RmpS(`+\^9%J:GHMub1Jk,l-rRADT<tFhjhcmC8_Vo>`\GsAs=<\E\)*tafK7_:h,%rg0OouMe"[Rs0"I*UFd'G4t5ClRA<A?@-Pujqgaq:]i:rsEZYtV$"k)_Rm2'O:hH:5j9]Nm:qgLeG^NaU)<Q)Gkf-E2s_`$\6@[#3f0+NTOMUY$[@D_l9NWgHP\m#OQlJu9=sHY#,liY`.X.RF!,Wf1%;_@7`>E'Hns4**7paL&m-7-:IH16JQ87%!!A/=RkgZ=]IiE5_$]s4?9#\CJ>;cH5C!%\I,aPPfdXm-qONd5Xp^0ZVt,WjH<Rm&6[gB>R(a=e"9E<]QG[)/80Z5[u(CA;KK7-1eeOZ^OZTQHVVI)<sEp/<t*pH5q5fcNbd"l4kk<QkQ#tE5?N#'LIn[U96T*8Z"-E>@DgrCZ*T#?Vt?Np/@*q[R$BGL2k$t7BHicji'\C9p!!)@kR'?IYWjLGg7m0dF"7?;+`YpZl=;_dO-J',uLXp0_!lo8;MMA(A/u4ko/>18/=ir-pGC]";uNm&Tu)p+R*$D1kgSX(Rljb0C'/Dlg"jjJOebU&kQo>oX_-'8ueW:hfJ+9;Ro'&;i)YM=XqcVj,A^6:%Jlk;=4%nkLP:DAZg<)#u=/.-_Lp/R3RT=LH-E6;P57oeZo(FEVh/UnB;2cT7cn(3JHIY3f':)7ZEJu).ARYUj0<L;e!U"T'RYY3kimREWrL%?(qX93:o1$hic"g!\:70QC>E>h+M%[T(EsLPOR:9EfDrBDG6cm77'N7V;,=e:0Muo<5\$3M@Rtgh"Hlc<Osg<H+F?p5)7:[5ts@MG2pqeje5_811tlL[75\/7nqg,3Gl2kW+ERj_+g'#hBC[_Rg<1?a:!,#1a3>jb3En63\P0[;7)%DCNK\3B;GV]hl^V:+HHiq/41E@lR!j-T.@+Z7s/[G_M%2E1H(5D.C4tj88IOJ7EVC"IR`RFoTM+M4%$;1jfX%H#GrtDV&O"7F(#po-4/7ah:E,fIAr02:C:$4Le>7gh%L^=P0[]..pA@?6.,a(45Ug6lU4rp&"pbX^Rcn:E<~>endstream
-endobj
-xref
-0 11
-0000000000 65535 f 
-0000000073 00000 n 
-0000000134 00000 n 
-0000000241 00000 n 
-0000000353 00000 n 
-0000000472 00000 n 
-0000000555 00000 n 
-0000000749 00000 n 
-0000000817 00000 n 
-0000001100 00000 n 
-0000001159 00000 n 
-trailer
-<<
-/ID 
-[<35ea57ffca9ecae590c1c4634fa948a2><35ea57ffca9ecae590c1c4634fa948a2>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 8 0 R
-/Root 7 0 R
-/Size 11
->>
-startxref
-2630
-%%EOF
