--- public/resources/lesson-plan-template-primary.pdf
@@ -1,254 +1,0 @@
-%PDF-1.4
-1 0 obj
-<< /Type /Catalog /Pages 2 0 R >>
-endobj
-2 0 obj
-<< /Type /Pages /Kids [3 0 R 4 0 R 5 0 R] /Count 3 >>
-endobj
-3 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 6 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-4 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 10 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-5 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 11 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-6 0 obj
-<< /Length 900 >>
-stream
-BT
-/F1 32 Tf
-0.424 0.388 1 rg
-50 720 Td
-(Primary School Lesson Plan Template) Tj
-0 0 0 rg
-/F2 18 Tf
-50 680 Td
-(Customizable template for elementary educators) Tj
-0.424 0.388 1 rg
-/F1 14 Tf
-50 640 Td
-(Streamline your planning with this comprehensive template) Tj
-0 0 0 rg
-/F2 12 Tf
-50 600 Td
-(This template includes sections for:) Tj
-0 -20 Td
-(• Learning objectives and standards alignment) Tj
-0 -20 Td
-(• Materials and resources needed) Tj
-0 -20 Td
-(• Step-by-step lesson activities) Tj
-0 -20 Td
-(• Differentiation strategies) Tj
-0 -20 Td
-(• Assessment and reflection notes) Tj
-0 -20 Td
-(• Extension activities for early finishers) Tj
-0 -30 Td
-/F1 14 Tf
-0.424 0.388 1 rg
-(Perfect for K-5 teachers planning engaging, standards-based lessons) Tj
-1 0.8 0.94 rg
-50 350 560 80 re f
-0 0 0 rg
-/F1 12 Tf
-60 410 Td
-(📚 This template is fully editable! Adapt sections to match your teaching style) Tj
-0 -15 Td
-(and curriculum requirements. Save time while maintaining quality planning.) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 1) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-7 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica-Bold >>
-endobj
-8 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
-endobj
-10 0 obj
-<< /Length 1100 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(Lesson Plan Template) Tj
-0 0 0 rg
-/F2 12 Tf
-50 720 Td
-(Teacher: _________________________ Grade Level: _____ Subject: _____________) Tj
-0 -15 Td
-(Date: ____________ Time: _________ Duration: _____ minutes) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(LEARNING OBJECTIVES) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Students will be able to:) Tj
-0 -15 Td
-(1. ________________________________________________________________) Tj
-0 -15 Td
-(2. ________________________________________________________________) Tj
-0 -15 Td
-(3. ________________________________________________________________) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(STANDARDS ALIGNMENT) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(________________________________________________________________) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(MATERIALS & RESOURCES) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• _____________________________________________________________) Tj
-0 -15 Td
-(• _____________________________________________________________) Tj
-0 -15 Td
-(• _____________________________________________________________) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(LESSON ACTIVITIES) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Opening (5 minutes): __________________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(Main Activity (25 minutes): ____________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(Closing (10 minutes): _________________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 2) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-11 0 obj
-<< /Length 1000 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 16 Tf
-50 750 Td
-(DIFFERENTIATION STRATEGIES) Tj
-/F2 12 Tf
-0 0 0 rg
-50 725 Td
-(For struggling learners: _______________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(For advanced learners: _______________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(For English Language Learners: ________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(ASSESSMENT) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(Formative Assessment: _______________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(Summative Assessment: ______________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(HOMEWORK/EXTENSION) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(________________________________________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(REFLECTION NOTES) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(What worked well? __________________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(What would I change next time? ________________________________) Tj
-0 -15 Td
-(________________________________________________________________) Tj
-0 -20 Td
-(Student engagement level: ____________________________________) Tj
-1 0.8 0.94 rg
-50 150 560 60 re f
-0 0 0 rg
-/F1 12 Tf
-60 190 Td
-(💡 Save this template and reuse it! Consistent planning saves time and) Tj
-0 -15 Td
-(ensures you never miss important lesson components.) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 3) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-xref
-0 12
-0000000000 65535 f 
-0000000009 00000 n 
-0000000058 00000 n 
-0000000127 00000 n 
-0000000260 00000 n 
-0000000393 00000 n 
-0000000526 00000 n 
-0000001478 00000 n 
-0000001537 00000 n 
-0000001593 00000 n 
-0000001633 00000 n 
-0000002785 00000 n 
-trailer
-<< /Size 12 /Root 1 0 R >>
-startxref
-3837
-%%EOF
