--- public/resources/lesson-plan-template-secondary.pdf
@@ -1,45 +1,0 @@
-%PDF-1.4
-1 0 obj
-<< /Type /Catalog /Pages 2 0 R >>
-endobj
-2 0 obj
-<< /Type /Pages /Kids [3 0 R] /Count 1 >>
-endobj
-3 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 4 0 R /Resources << /Font << /F1 5 0 R >> >> >>
-endobj
-4 0 obj
-<< /Length 180 >>
-stream
-BT
-/F1 18 Tf
-50 750 Td
-(Secondary Lesson Plan Template (Editable)) Tj
-0 -30 Td
-/F1 12 Tf
-(Placeholder content for secondary lesson template) Tj
-0 -20 Td
-(This PDF will contain customizable templates) Tj
-0 -20 Td
-(for secondary school lesson planning.) Tj
-0 -40 Td
-(Zaza Technologies - www.zazateach.com) Tj
-ET
-endstream
-endobj
-5 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
-endobj
-xref
-0 6
-0000000000 65535 f 
-0000000009 00000 n 
-0000000058 00000 n 
-0000000115 00000 n 
-0000000274 00000 n 
-0000000506 00000 n 
-trailer
-<< /Size 6 /Root 1 0 R >>
-startxref
-573
-%%EOF
