--- public/resources/management/behavior-redirection-scripts-v1.0-en.pdf
@@ -1,80 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 9 0 R /MediaBox [ 0 0 612 792 ] /Parent 8 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/PageMode /UseNone /Pages 8 0 R /Type /Catalog
->>
-endobj
-7 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-8 0 obj
-<<
-/Count 1 /Kids [ 5 0 R ] /Type /Pages
->>
-endobj
-9 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 824
->>
-stream
-Gatn#;/=o?&BE]*.;jRdZE/%_78BB@Be9],$=dr*?4''-c`nDCE[A81#l3mMa\AI5n=V=e(nY?rn,E(571dl^q>[f'3+'Jc4G&C*,):sZ*#;WFF`I%m'KtoT-8:iBg.X6T*PHi.7.4*jR![k8c@Bp/N?I<BBKW-g1lRV$Jr%Bn#'pT.F"G8)<#TXP/K8K`4^;JD25XESqBIQZ@52lG^XMioJFa6noJ+%,9bL/?am"3N;MSa#>_o[HJQ1/h)N2"W;Uilj/.loS86&Zh)U_M#UQqjP;AF%(n-O,4a`=]m)W[nLeU#_'-Kp.o'*n>71cu>/P=\hP^lo(o`Xjg^j`(F>N6O=^nRm\e=!<2EnJ*Cb_TR8dq8%01p>NuS8Y:a=90cM`I0/<W_/2$"_Jj`gjd5au-)?7!(A6@"#ff/o9[SA5@Js@D^h!upeUGjjZ)jcN0KT;FKsk@sKGpI64^o'EpQEP0h<lTfQKKJCbfU#ob@P+U<bn<JT,Ma=E*`W\i"u=n\1)+!_EG:rO_s#oe#i=BEqX+fCWof4)RuPi*U5tm&F<Rd-0s-*W]FgQ#"c-jg!B)sVSKm.22F9qX#D/e6Iu/:*?Xi`*4#V'^puZZUp3?T:UM*DXI<$`$%5K0H*_`3=Q5nFD*0'%[bDbUAor!94"tO#-uC[@d@KT\35DOfpmuB4_G?WVmZqTnUMG+L<L^\feW7aBn*o:!Yk.hWE2Z!,3erfC_M;0hX60[#KpB&UK`]F%f$Y>l4VGOH$O)Ne'50uP``_:uU:9d<K^uPl=[e*FBUWW=VQ5l7?GA4>bRPJBNt4_S_[VWT!uTX^_u~>endstream
-endobj
-xref
-0 10
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000655 00000 n 
-0000000723 00000 n 
-0000001006 00000 n 
-0000001065 00000 n 
-trailer
-<<
-/ID 
-[<95e5753d5eb57bfbec80c248d31d3a70><95e5753d5eb57bfbec80c248d31d3a70>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 7 0 R
-/Root 6 0 R
-/Size 10
->>
-startxref
-1979
-%%EOF
