--- public/resources/pdfs/demo-resource.pdf
@@ -1,69 +1,0 @@
-%PDF-1.4
-1 0 obj
-<<
-/Type /Catalog
-/Pages 2 0 R
->>
-endobj
-
-2 0 obj
-<<
-/Type /Pages
-/Kids [3 0 R]
-/Count 1
->>
-endobj
-
-3 0 obj
-<<
-/Type /Page
-/Parent 2 0 R
-/MediaBox [0 0 612 792]
-/Resources <<
-/Font <<
-/F1 4 0 R
->>
->>
-/Contents 5 0 R
->>
-endobj
-
-4 0 obj
-<<
-/Type /Font
-/Subtype /Type1
-/BaseFont /Helvetica
->>
-endobj
-
-5 0 obj
-<<
-/Length 100
->>
-stream
-BT
-/F1 12 Tf
-72 720 Td
-(Zaza Promptly - Sample Resource) Tj
-0 -24 Td
-(This is a demonstration of our resource system.) Tj
-ET
-endstream
-endobj
-
-xref
-0 6
-0000000000 65535 f 
-0000000010 00000 n 
-0000000053 00000 n 
-0000000110 00000 n 
-0000000250 00000 n 
-0000000317 00000 n 
-trailer
-<<
-/Size 6
-/Root 1 0 R
->>
-startxref
-467
-%%EOF
