--- public/resources/planning/differentiated-question-stems-v1.0-en.pdf
@@ -1,80 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 9 0 R /MediaBox [ 0 0 612 792 ] /Parent 8 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/PageMode /UseNone /Pages 8 0 R /Type /Catalog
->>
-endobj
-7 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-8 0 obj
-<<
-/Count 1 /Kids [ 5 0 R ] /Type /Pages
->>
-endobj
-9 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1005
->>
-stream
-Gatn%9i'M/&;KZOM@h4lMXe_JES<e!=!]=G8e]\^Q]4."(2U2->pgK-J+)_b+I/fNMNDDm)Hf/J^,,YH"7YO^gbUM^NlQ,6/V"ssnIu&e0c495fm)\/-WkdW8>&0Ul!oTr5)Y#+!1lFEj!4XY#\,aq2;=gIK%Hh0BXNHC%3]8*h(Z4:m[]7"'GZp'E.]$5eYnIh-&`6JH6:/90^dWL=*4D35>Kr1^5K[!\&oK"m_WNscD<'%8!b$YTM:Up4XggkKpdSU77@g+opl:nAoMi%Q'+RI[+>W$XD7c',IT5iq^^&rdLQR;G[WR>E<+iL*&9j/]++d/ZljAC8DH+n3,l3Z%usfMlurKMM8+U><,^ZjB^9[P$jP)ErPAVo$*)=b*aAtf5!ap%;;rs"R.<B6TXS^s@-.`=fKdE@d3,>GTc`as3jJON'Q?4dFd\rOL8ne#7WhRI,p%mM+0(ihdUJcTYHXNU)o0DjX^R6i.W6lY[3br,o*&<c$<_,=W2hMoBJ,WZmmbm4P]`R7^r4n!9I8#S8$l5CeA>H&;G&Jrg$4gEP1[.HCqo/g:u.'r<C/#Q@O[:-E<F3/gm[bL_=O+_/EHOdRh=4g9rEKpZrA6?6\/i%1CtH38R7'](9%lRF9e$[qt1uBE;S3N;7M3>74i,J's4F1G4fZ=6ifEZ\o@sD[BH-J)7b<?_)d?ZabMjoQs>ToN=I2('68t%"$%tfmrf&hhd3H_>uac?X4Il"W-"8l9ZkH-0ne(GJ"\3jS3e2N_.SWA/+]SDJ7XVAbak'Cf_3m9,*a\DGUX0e>Zc_qhYMu/,,%">=X's6[2UFGhDie%B[9i9@2%o`QmdMaONPih,>D*F$^3&+RL&m&agsJH%Jq6p:G`H/\lmZS&G8*obKf&GM6T.BB/:J)&B"J3(P%7E!U$S?.lV6J)#-MM#+hEFX@2p_$r[e?&.b-Kn;55?BrHmRN#G<8E*fD%^"[SAOOM])fem@1Fusocet(,89%k][quK&N=)%~>endstream
-endobj
-xref
-0 10
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000655 00000 n 
-0000000723 00000 n 
-0000001006 00000 n 
-0000001065 00000 n 
-trailer
-<<
-/ID 
-[<7e0f42e1bf85f863a451a33d7eb6b759><7e0f42e1bf85f863a451a33d7eb6b759>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 7 0 R
-/Root 6 0 R
-/Size 10
->>
-startxref
-2161
-%%EOF
