--- public/resources/planning/lesson-plan-template-primary-v1.0-en.pdf
@@ -1,80 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/Contents 9 0 R /MediaBox [ 0 0 612 792 ] /Parent 8 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-6 0 obj
-<<
-/PageMode /UseNone /Pages 8 0 R /Type /Catalog
->>
-endobj
-7 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-8 0 obj
-<<
-/Count 1 /Kids [ 5 0 R ] /Type /Pages
->>
-endobj
-9 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 786
->>
-stream
-Gatn$gMY_1&;KY!MG^MVKoL;i50uA0V(#<BUl)"K&i[fO@"V8-l_GrO]4Ypi<>*n3'NlE_ZK@X+k2.$d#!KWu%Ussl*aF*Q^]es=i0$\$DW&\'X?(b+VW37k'_F/O0ghDbFFY6h@C\!E/hB#9I,:PVHlmZ"d:.t`5fbFX&676J?*7X$?c*<g7@M"tj<rkoS4103*dA4RodfZ1HJ[YXgu+1glog?_)k+<=;:BCr*+]$g+kkk[$MU@g!n-<8#$O0!E=TedK4P(i1l!gm0e]o+KIWC^"9$)M%";'P6LW@g-E5":[:UrU.bbK37`24eg=sJ:f+sAM9F+-l*8"'r--:G9e^tV&]<tVIQ9*IP_jp?&#a=[M7F@p_QD.sFD$Aa=O'H,COA5Hrgi9"E^G"p<?$L(\JhR(O/&uCA>V"u\f-F6nkVC(E%q`Ah,`L<<hBB*;ruj'%a>%Y%.uh+1&#%4EG$B\\7:Ye+<?cRn/HP)WTp7G/Su,/\XdBB&>7dj2h(t3X+*G_9S0SsL84/>l5)(aS^pQTEf.oP92&TYc.`;n(ga@$j\)eDI:-geG=IMaI>;@&ER*!FhA8&j[:<UP&ZY=SRmXCaG0t@(T1m%JYl;f="9i.V,"f<M?iV(HEbi6&fc(ddf[j^8cD#eUj)E]V8Z;4C`&*_"Pd&eh*U8>+EWDK'1d5pnofiT&>P.K!U)SOfj9c6dC,7AL7N="aSK>@]gCC\u3T`AFeW+Ij0/'h8r<)?&19q1efUJ8XtTD)R_jb985B'M359A%u+6K])5j99l~>endstream
-endobj
-xref
-0 10
-0000000000 65535 f 
-0000000073 00000 n 
-0000000124 00000 n 
-0000000231 00000 n 
-0000000343 00000 n 
-0000000462 00000 n 
-0000000655 00000 n 
-0000000723 00000 n 
-0000001006 00000 n 
-0000001065 00000 n 
-trailer
-<<
-/ID 
-[<75642609bc297c549400ac99cb837106><75642609bc297c549400ac99cb837106>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 7 0 R
-/Root 6 0 R
-/Size 10
->>
-startxref
-1941
-%%EOF
