--- public/resources/planning/lesson-plan-template-secondary-v1.0-en.pdf
@@ -1,86 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R /F4 5 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Helvetica-BoldOblique /Encoding /WinAnsiEncoding /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/BaseFont /Symbol /Name /F4 /Subtype /Type1 /Type /Font
->>
-endobj
-6 0 obj
-<<
-/Contents 10 0 R /MediaBox [ 0 0 612 792 ] /Parent 9 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-7 0 obj
-<<
-/PageMode /UseNone /Pages 9 0 R /Type /Catalog
->>
-endobj
-8 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250915083035+02'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250915083035+02'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-9 0 obj
-<<
-/Count 1 /Kids [ 6 0 R ] /Type /Pages
->>
-endobj
-10 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 721
->>
-stream
-Gatm99lHLd&A@sBN,kk5X22EknmUG2`fVaID5)"AKu2hc""[jsfhn"6h>uY9Fssb3KnNi9k.Q'37=jAcRsUA#HjX8)9n>&*&tn%>&DY&N(d3jlHd_N%SMNBe2N)ODg)aA7,ED5F7@D'V:i8bdLVC=56uNt>A\`QXQa,91pg-?<O2LN+DM6RjLYFZ0+V7e=L.4iZp^+dD#/A(AR"45c$'</MMV#_uDdbJAbDecob>cDVe3I5P*l,ijjr=7219$C1,RPbrNpgBt6S+EhPTi%AXK>MV-!guSc>)r'.j;$Dgio1LJAmnr*=CAfbrLqN%>Z"rUOEl<hZWSb19&KEKlVQ]L/=7UdeMY]APG;uVUI%!hp`Fm@qRr-9W,QXmQ(+,L'^6b4Wk(XItr%!_(`?0(`=1t6t0dOEBY4R=VU,lR/tuW0[+8Zc$@E>DFY0,6P$pb[!d5u0"4&H.?)"_[1.M6;Wn1SK^PVR]qKEfiB`)N]B1IGRH(__:0bS?-)3td7A`>8&?.-<f\j@Q7uleMDshe+Gf\dW2=f;fSE]-$LFRLT'nn-^!d#,'M+V9JK5K\\5BJZ&L9,P&8+$FNM&foel:3>07e]e!q,I#;DU5paVsXJMdCK/V9<G$K]aX\_^b7m<!`p5+XhkO"WPt-DQp4akNI@l%l'R;sU\+a&'^"MrW'WtWYP[Fg<TPJ)W`<4[m:BSqdUc]'aJQ(u0_tsL!Sa(~>endstream
-endobj
-xref
-0 11
-0000000000 65535 f 
-0000000073 00000 n 
-0000000134 00000 n 
-0000000241 00000 n 
-0000000353 00000 n 
-0000000472 00000 n 
-0000000549 00000 n 
-0000000743 00000 n 
-0000000811 00000 n 
-0000001094 00000 n 
-0000001153 00000 n 
-trailer
-<<
-/ID 
-[<2b98f749da4f75c85b16f99738d1b696><2b98f749da4f75c85b16f99738d1b696>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 8 0 R
-/Root 7 0 R
-/Size 11
->>
-startxref
-1965
-%%EOF
