--- public/resources/teacher-ai-toolkit.pdf
@@ -1,204 +1,0 @@
-%PDF-1.4
-1 0 obj
-<< /Type /Catalog /Pages 2 0 R >>
-endobj
-2 0 obj
-<< /Type /Pages /Kids [3 0 R 4 0 R 5 0 R] /Count 3 >>
-endobj
-3 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 6 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> /ExtGState << /GS1 9 0 R >> >> >>
-endobj
-4 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 10 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-5 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 11 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-6 0 obj
-<< /Length 800 >>
-stream
-BT
-/F1 32 Tf
-0.424 0.388 1 rg
-50 720 Td
-(The Essential AI Prompt Toolkit for Teachers) Tj
-0 0 0 rg
-/F2 18 Tf
-50 680 Td
-(A comprehensive collection of AI prompts to transform your teaching) Tj
-0.424 0.388 1 rg
-/F1 14 Tf
-50 640 Td
-(Save hours on lesson planning, grading, and classroom management) Tj
-0 0 0 rg
-/F2 12 Tf
-50 600 Td
-(This resource includes:) Tj
-0 -20 Td
-(• 50+ proven AI prompts for lesson planning) Tj
-0 -20 Td
-(• Templates for differentiated instruction) Tj
-0 -20 Td
-(• Time-saving grading prompts) Tj
-0 -20 Td
-(• Parent communication templates) Tj
-0 -20 Td
-(• Classroom management strategies) Tj
-0.424 0.388 1 rg
-/F1 16 Tf
-50 450 Td
-(💡 Tip: Start with the basic prompts, then customize for your classroom!) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 1) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-7 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica-Bold >>
-endobj
-8 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
-endobj
-9 0 obj
-<< /Type /ExtGState /ca 0.3 >>
-endobj
-10 0 obj
-<< /Length 900 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(Quick Start Guide) Tj
-0 0 0 rg
-/F2 12 Tf
-50 720 Td
-(Getting started with AI prompts is easier than you think. Follow these simple steps:) Tj
-0 -30 Td
-/F1 14 Tf
-(Step 1: Choose Your Tool) Tj
-/F2 12 Tf
-0 -20 Td
-(Popular options include ChatGPT, Claude, or Google Bard. Each has strengths for) Tj
-0 -15 Td
-(different educational tasks.) Tj
-0 -30 Td
-/F1 14 Tf
-(Step 2: Start with Basic Prompts) Tj
-/F2 12 Tf
-0 -20 Td
-(Begin with simple requests like "Create a lesson plan for 4th grade math on fractions") Tj
-0 -15 Td
-(that includes hands-on activities." Build complexity as you get comfortable.) Tj
-0 -30 Td
-/F1 14 Tf
-(Step 3: Customize for Your Students) Tj
-/F2 12 Tf
-0 -20 Td
-(Always add context about your specific students, classroom size, and learning needs.) Tj
-0 -15 Td
-(The more specific you are, the better the AI can help you.) Tj
-1 0.8 0.94 rg
-50 350 560 80 re f
-0 0 0 rg
-/F1 12 Tf
-60 410 Td
-(💡 Pro Tip: Save your best prompts in a document so you can reuse and refine them) Tj
-0 -15 Td
-(over time. Great prompts are like great lesson plans - they improve with iteration!) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 2) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-11 0 obj
-<< /Length 950 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(Sample Prompts to Get Started) Tj
-0 0 0 rg
-/F1 16 Tf
-50 720 Td
-(Lesson Planning Prompts) Tj
-/F2 12 Tf
-50 695 Td
-("Create a 45-minute science lesson for 5th graders on the water cycle that includes) Tj
-0 -15 Td
-(a hands-on experiment, visual aids, and assessment questions.") Tj
-0 -30 Td
-/F1 16 Tf
-(Grading & Feedback Prompts) Tj
-/F2 12 Tf
-0 -20 Td
-("Provide constructive feedback on this student essay about friendship. Focus on) Tj
-0 -15 Td
-(strengths first, then suggest specific improvements for organization and voice.") Tj
-0 -30 Td
-/F1 16 Tf
-(Parent Communication Prompts) Tj
-/F2 12 Tf
-0 -20 Td
-("Draft a positive email to parents about their child's improvement in reading) Tj
-0 -15 Td
-(comprehension. Include specific examples and suggest ways to support at home.") Tj
-0 -30 Td
-/F1 16 Tf
-(Differentiation Prompts) Tj
-/F2 12 Tf
-0 -20 Td
-("Adapt this math worksheet on multiplication for three levels: below grade level,) Tj
-0 -15 Td
-(on grade level, and advanced. Maintain the same learning objectives.") Tj
-1 0.8 0.94 rg
-50 250 560 60 re f
-0 0 0 rg
-/F1 12 Tf
-60 290 Td
-(📧 Want more prompts? Visit www.zazateach.com for our complete prompt library) Tj
-0 -15 Td
-(and join thousands of teachers transforming their practice with AI!) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 3) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-xref
-0 12
-0000000000 65535 f 
-0000000009 00000 n 
-0000000058 00000 n 
-0000000127 00000 n 
-0000000295 00000 n 
-0000000428 00000 n 
-0000000561 00000 n 
-0000001413 00000 n 
-0000001472 00000 n 
-0000001528 00000 n 
-0000001568 00000 n 
-0000002520 00000 n 
-trailer
-<< /Size 12 /Root 1 0 R >>
-startxref
-3522
-%%EOF
