--- public/resources/teacher-self-care-guide.pdf
@@ -1,196 +1,0 @@
-%PDF-1.4
-1 0 obj
-<< /Type /Catalog /Pages 2 0 R >>
-endobj
-2 0 obj
-<< /Type /Pages /Kids [3 0 R 4 0 R] /Count 2 >>
-endobj
-3 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 6 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-4 0 obj
-<< /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 10 0 R /Resources << /Font << /F1 7 0 R /F2 8 0 R >> >> >>
-endobj
-6 0 obj
-<< /Length 950 >>
-stream
-BT
-/F1 32 Tf
-0.424 0.388 1 rg
-50 720 Td
-(Teacher Self-Care Guide: Quick Wins) Tj
-0 0 0 rg
-/F2 18 Tf
-50 680 Td
-(Practical self-care strategies that fit into a teacher's busy schedule) Tj
-0.424 0.388 1 rg
-/F1 14 Tf
-50 640 Td
-(You can't pour from an empty cup - prioritize your wellbeing) Tj
-0 0 0 rg
-/F2 12 Tf
-50 600 Td
-(Teaching is rewarding but demanding. This guide provides realistic, research-based) Tj
-0 -15 Td
-(self-care strategies that busy educators can actually implement:) Tj
-0 -30 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(5-Minute Morning Reset) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• Take 3 deep breaths before entering school) Tj
-0 -15 Td
-(• Set one positive intention for the day) Tj
-0 -15 Td
-(• Drink a full glass of water) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Lunch Break Boundaries) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• Eat away from your classroom) Tj
-0 -15 Td
-(• Take a 10-minute walk outside) Tj
-0 -15 Td
-(• Connect with a colleague for positive conversation) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(After-School Wind Down) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• Leave work at work - set a firm cutoff time) Tj
-0 -15 Td
-(• Do something just for you for 20 minutes) Tj
-0 -15 Td
-(• Practice gratitude - name 3 good things from the day) Tj
-1 0.8 0.94 rg
-50 200 560 80 re f
-0 0 0 rg
-/F1 12 Tf
-60 260 Td
-(💜 Remember: Self-care isn't selfish - it's essential! When you take care of) Tj
-0 -15 Td
-(yourself, you're better able to take care of your students.) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 1) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-7 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica-Bold >>
-endobj
-8 0 obj
-<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica >>
-endobj
-10 0 obj
-<< /Length 1100 >>
-stream
-BT
-0.424 0.388 1 rg
-/F1 24 Tf
-50 750 Td
-(Sustainable Self-Care Strategies) Tj
-0 0 0 rg
-/F1 16 Tf
-50 720 Td
-(Physical Wellbeing) Tj
-/F2 12 Tf
-0 -20 Td
-(• Pack healthy snacks - avoid the vending machine trap) Tj
-0 -15 Td
-(• Use a water bottle with time markers to stay hydrated) Tj
-0 -15 Td
-(• Take stairs when possible for micro-workouts) Tj
-0 -15 Td
-(• Stand and stretch every 30 minutes during planning) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Mental Wellbeing) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• Practice the 2-minute rule: if it takes less than 2 minutes, do it now) Tj
-0 -15 Td
-(• Use positive self-talk - you're doing important work) Tj
-0 -15 Td
-(• Keep a victory journal of teaching wins, big and small) Tj
-0 -15 Td
-(• Set boundaries with technology - no emails after 7 PM) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Emotional Wellbeing) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• Find your teacher tribe - connect with supportive colleagues) Tj
-0 -15 Td
-(• Remember your 'why' - you're making a difference) Tj
-0 -15 Td
-(• Practice saying no to non-essential commitments) Tj
-0 -15 Td
-(• Celebrate progress, not perfection) Tj
-0 -25 Td
-/F1 16 Tf
-0.424 0.388 1 rg
-(Weekend Recovery) Tj
-/F2 12 Tf
-0 0 0 rg
-0 -20 Td
-(• Dedicate one day to non-school activities) Tj
-0 -15 Td
-(• Engage in a hobby that brings you joy) Tj
-0 -15 Td
-(• Get outside - nature is a natural stress reliever) Tj
-0 -15 Td
-(• Connect with friends and family outside of education) Tj
-1 0.8 0.94 rg
-50 120 560 60 re f
-0 0 0 rg
-/F1 12 Tf
-60 160 Td
-(🌱 Start small! Pick ONE strategy and commit to it for two weeks.) Tj
-0 -15 Td
-(Small, consistent changes create lasting wellbeing habits.) Tj
-0.5 0.5 0.5 rg
-/F2 10 Tf
-50 50 Td
-(Page 2) Tj
-306 50 Td
-(Zaza Technologies · www.zazateach.com) Tj
-500 50 Td
-([ZAZA]) Tj
-ET
-endstream
-endobj
-xref
-0 11
-0000000000 65535 f 
-0000000009 00000 n 
-0000000058 00000 n 
-0000000120 00000 n 
-0000000253 00000 n 
-0000000386 00000 n 
-0000000519 00000 n 
-0000001521 00000 n 
-0000001580 00000 n 
-0000001636 00000 n 
-0000001676 00000 n 
-trailer
-<< /Size 11 /Root 1 0 R >>
-startxref
-2828
-%%EOF
