--- public/resources/teacher-self-care-guide/teacher-self-care-guide-v2.md
@@ -1,167 +1,0 @@
-# Teacher Self-Care Guide: Reclaim Your Time and Energy
-
-*You're not just a teacher—you're a human being who deserves to thrive, not just survive.*
-
----
-
-## Why This Guide Matters
-
-Teaching is one of the most demanding professions, yet self-care often feels like another item on an impossible to-do list. This guide isn't about bubble baths and meditation apps (though those can help). It's about creating realistic, sustainable practices that fit into your actual teaching life.
-
-> "I used to grade papers until 11 PM every night. Now I have systems that let me finish by 7 PM and actually have dinner with my family." — Sarah M., 4th grade teacher
-
-## Quick Wins: Start Here Today
-
-### 🕒 Time Boundaries That Actually Work
-
-**The 3-2-1 Rule**
-- **3 PM:** Stop checking work emails  
-- **2 hours:** Maximum evening work time
-- **1 day:** Complete work-free weekend day
-
-**Micro-Breaks During the School Day**
-- **Between classes:** 30 seconds of deep breathing
-- **Lunch time:** Eat away from your desk (even if it's just the hallway)
-- **Planning period:** 10 minutes outside or by a window
-
-### 📧 Email Sanity Savers
-
-**Parent Email Stress-Busters:**
-- Use templates for common responses (save 15+ minutes daily)
-- Set auto-responses: "I'll respond within 24 hours during school days"
-- Batch process emails at set times: morning, lunch, after school
-
-*"I stopped checking emails at home and my anxiety dropped dramatically. Parents understand—they don't need instant responses."* — Marcus T., high school English
-
-### 🏃‍♀️ Physical Energy Management
-
-**Classroom Movement Hacks:**
-- Walk the room while students work independently
-- Use a standing desk converter ($30 solution)
-- Take stairs instead of elevator when possible
-- Park further away for extra steps
-
-**Energy-Boosting Snacks to Keep in Desk:**
-- Mixed nuts and dried fruit
-- Apple slices with peanut butter packets  
-- Herbal tea bags for afternoon slump
-- Dark chocolate (really—it helps focus!)
-
----
-
-## Sustainable Systems for Long-Term Success
-
-### Morning Routines for Teaching Success
-
-**The 15-Minute Morning Reset**
-1. **5 minutes:** Review today's schedule, adjust if needed
-2. **5 minutes:** Prepare classroom while drinking coffee/tea  
-3. **5 minutes:** Quick positive intention setting
-
-**Sunday Prep That Prevents Monday Meltdowns:**
-- Layout clothes for the week
-- Prep 3 days of lesson materials  
-- Batch cook easy lunches
-- Set up workspace for productive start
-
-### Grading Without the Guilt
-
-**The Reality Check:** You don't need to grade everything perfectly.
-
-**Smart Grading Strategies:**
-- **Priority system:** Grade formative assessments lightly, summative thoroughly
-- **Voice memos:** Record feedback while walking—faster than writing
-- **Student self-assessment:** Let them do initial evaluation, you verify
-- **Focus on growth:** Note improvement trends rather than perfection
-
-### Difficult Parent Communication
-
-**Before the Conversation:**
-- Write key points on paper (prevents rambling)
-- Have student work samples ready
-- Set time limits: "I have 15 minutes to discuss..."
-
-**Stress-Reducing Phrases:**
-- "I understand your concern. Here's what I'm seeing..."
-- "Let's work together to support [student name]..."  
-- "I'll follow up with you by [specific date]"
-
----
-
-## Your 4-Week Action Plan
-
-### Week 1: Foundation
-- [ ] Set email boundaries (3-2-1 rule)
-- [ ] Create one template for common parent emails
-- [ ] Identify one 10-minute daily self-care activity
-
-### Week 2: Systems  
-- [ ] Implement smart grading strategy for one subject
-- [ ] Set up Sunday prep routine
-- [ ] Create classroom snack stash
-
-### Week 3: Sustainability
-- [ ] Evaluate what's working, adjust what isn't
-- [ ] Share successful strategies with colleague
-- [ ] Plan one work-free weekend day
-
-### Week 4: Growth
-- [ ] Add second self-care practice
-- [ ] Refine morning routine
-- [ ] Celebrate progress (this matters!)
-
----
-
-## Quick Reference Card
-*Print and keep in your desk drawer*
-
-### Daily Check-In Questions
-1. What's one thing that went well today?
-2. What's one thing I can improve tomorrow?
-3. Did I take care of my basic needs (food, hydration, movement)?
-
-### Emergency Stress Response
-**When overwhelmed:**
-1. Stop, breathe deeply (4 counts in, 6 counts out)
-2. Write down three things you can control right now
-3. Do one small thing to help future you
-4. Remember: This feeling will pass
-
-### Support Network Reminders
-- **Trusted colleague** for venting: _____________
-- **Admin to contact** when overwhelmed: _____________  
-- **Professional support** if needed: _____________
-
----
-
-## Remember This
-
-Self-care isn't selfish—it's essential. You can't pour from an empty cup. By taking care of yourself, you're modeling healthy boundaries for your students and ensuring you can teach with energy and joy for years to come.
-
-*Your students need you healthy, happy, and sustainable.*
-
----
-
-## Research Foundation
-
-This guide is based on research from the National Education Association showing that teacher stress primarily comes from:
-1. **Workload intensity** (70% of stress)
-2. **Lack of control** (23% of stress)  
-3. **Difficult relationships** (7% of stress)
-
-All strategies in this guide address these core stress factors with practical, classroom-tested solutions.
-
----
-
-## About This Resource
-
-Created by educators, for educators. Real strategies from teachers across all grade levels who have successfully implemented sustainable self-care practices while maintaining teaching excellence.
-
-**For more teacher resources and time-saving tools, visit:**  
-**https://zazapromptly.com/resources**
-
-*Zaza Promptly - AI tools that give teachers their time back*
-
----
-
-*Version 2.0 | Updated January 2025 | © Zaza Technologies*
