--- public/resources/video-guide-series.pdf
@@ -1,112 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Video Guide Series - Zaza Technologies</title>
-    <style>
-        body {
-            font-family: system-ui, -apple-system, sans-serif;
-            max-width: 800px;
-            margin: 0 auto;
-            padding: 20px;
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            color: white;
-            min-height: 100vh;
-        }
-        .container {
-            background: rgba(255, 255, 255, 0.1);
-            padding: 30px;
-            border-radius: 20px;
-            backdrop-filter: blur(10px);
-        }
-        h1 {
-            font-size: 2.5rem;
-            margin-bottom: 1rem;
-            text-align: center;
-        }
-        .subtitle {
-            text-align: center;
-            font-size: 1.2rem;
-            margin-bottom: 2rem;
-            opacity: 0.9;
-        }
-        .video-list {
-            list-style: none;
-            padding: 0;
-        }
-        .video-item {
-            background: rgba(255, 255, 255, 0.1);
-            margin: 1rem 0;
-            padding: 1.5rem;
-            border-radius: 10px;
-            border-left: 4px solid #4CAF50;
-        }
-        .video-title {
-            font-size: 1.3rem;
-            font-weight: bold;
-            margin-bottom: 0.5rem;
-        }
-        .video-desc {
-            opacity: 0.8;
-            margin-bottom: 1rem;
-        }
-        .coming-soon {
-            background: #FF9800;
-            color: white;
-            padding: 0.3rem 0.8rem;
-            border-radius: 15px;
-            font-size: 0.8rem;
-            font-weight: bold;
-        }
-        .back-link {
-            display: inline-block;
-            margin-top: 2rem;
-            padding: 0.8rem 1.5rem;
-            background: rgba(255, 255, 255, 0.2);
-            color: white;
-            text-decoration: none;
-            border-radius: 25px;
-            transition: all 0.3s ease;
-        }
-        .back-link:hover {
-            background: rgba(255, 255, 255, 0.3);
-            transform: translateY(-2px);
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>🎥 Video Guide Series</h1>
-        <p class="subtitle">Step-by-step tutorials on using AI tools effectively in your classroom</p>
-        
-        <ul class="video-list">
-            <li class="video-item">
-                <div class="video-title">Getting Started with AI Tools</div>
-                <div class="video-desc">Learn the basics of integrating AI into your teaching workflow</div>
-                <span class="coming-soon">Coming Soon</span>
-            </li>
-            
-            <li class="video-item">
-                <div class="video-title">Effective AI Prompt Writing</div>
-                <div class="video-desc">Master the art of writing prompts that generate useful educational content</div>
-                <span class="coming-soon">Coming Soon</span>
-            </li>
-            
-            <li class="video-item">
-                <div class="video-title">Student Privacy and AI Ethics</div>
-                <div class="video-desc">Best practices for using AI while protecting student data</div>
-                <span class="coming-soon">Coming Soon</span>
-            </li>
-            
-            <li class="video-item">
-                <div class="video-title">Advanced AI Features</div>
-                <div class="video-desc">Unlock the full potential of AI tools for lesson planning and assessment</div>
-                <span class="coming-soon">Coming Soon</span>
-            </li>
-        </ul>
-        
-        <a href="javascript:history.back()" class="back-link">← Back to Resources</a>
-    </div>
-</body>
-</html>
